+++ packages/perseus/src/__stories__/article-renderer.stories.tsx
@@ -1,3 +1,4 @@
+import {KeypadContext} from "@khanacademy/math-input";
 import React from "react";
 
 import {storybookDependenciesV2} from "../../../../testing/test-dependencies";
@@ -5,8 +6,10 @@ import {
     singleSectionArticle,
     multiSectionArticle,
     passageArticle,
+    articleWithExpression,
 } from "../__testdata__/article-renderer.testdata";
 import ArticleRenderer from "../article-renderer";
+import TestKeypadContextWrapper from "../widgets/__stories__/test-keypad-context-wrapper";
 
 export default {
     title: "Perseus/Renderers/Article Renderer",
@@ -47,4 +50,23 @@ export const PassageArticle = ({useNewStyles}): any => (
         dependencies={storybookDependenciesV2}
         useNewStyles={useNewStyles}
     />
+);
+
+export const ExpressionArticle = ({useNewStyles}): any => (
+    <TestKeypadContextWrapper>
+        <KeypadContext.Consumer>
+            {({keypadElement, setRenderer, scrollableElement}) => (
+                <ArticleRenderer
+                    ref={(node) => {
+                        setRenderer(node);
+                    }}
+                    json={articleWithExpression}
+                    dependencies={storybookDependenciesV2}
+                    useNewStyles={useNewStyles}
+                    apiOptions={{isMobile: true, customKeypad: true}}
+                    keypadElement={keypadElement}
+                />
+            )}
+        </KeypadContext.Consumer>
+    </TestKeypadContextWrapper>
 );