+++ packages/perseus/src/__testdata__/article-renderer.testdata.ts
@@ -58,6 +58,34 @@ export const passageArticle: PerseusRenderer = {
     },
 };
 
+export const articleWithExpression: PerseusRenderer = {
+    content:
+        "### Practice Problem\n\n$8\\cdot(11i+2)=$ [[☃ expression 1]]  \n*Your answer should be a complex number in the form $a+bi$ where $a$ and $b$ are real numbers.*",
+    images: {},
+    widgets: {
+        "expression 1": {
+            alignment: "default",
+            graded: true,
+            options: {
+                answerForms: [
+                    {
+                        considered: "correct",
+                        form: true,
+                        simplify: false,
+                        value: "16+88i",
+                    },
+                ],
+                buttonSets: ["basic"],
+                functions: ["f", "g", "h"],
+                times: false,
+            },
+            static: false,
+            type: "expression",
+            version: {major: 1, minor: 0},
+        },
+    },
+};
+
 export const multiSectionArticle: ReadonlyArray<PerseusRenderer> = [
     {
         content: