+++ packages/perseus/src/types.ts
@@ -8,6 +8,7 @@ import type {AnalyticsEventHandlerFn} from "@khanacademy/perseus-core";
 import type {LinterContextProps} from "@khanacademy/perseus-linter";
 import type {Result} from "@khanacademy/wonder-blocks-data";
 import type * as React from "react";
+import type ReactDOM from "react-dom";
 
 export type FocusPath = ReadonlyArray<string> | null | undefined;
 
@@ -143,10 +144,7 @@ export type APIOptions = Readonly<{
     onFocusChange?: (
         newFocusPath: FocusPath,
         oldFocusPath: FocusPath,
-        // These two parameters are only used in library-exercise.jsx and
-        // practice-exercise.jsx and in the mobile app
-        // (javascript/perseus-package/perseus-mobile-bridge.js).
-        keypadElement?: HTMLElement,
+        keypadElement?: ReturnType<typeof ReactDOM.findDOMNode>,
         focusedElement?: HTMLElement,
     ) => unknown;
     GroupMetadataEditor?: React.ComponentType<StubTagEditorType>;